Move authorization handling of APIs to intercepter level
